[#68811] Move i18n.ts to services
https://community.openproject.org/wp/68811
