+++ lib/index.ts
@@ -1,3 +1,3 @@
-import "./i18n";
+import "./services/i18n.ts";
 export * from "./components";
 export { initializeOpBlockNoteExtensions } from "./initialize"
